health/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-babeljs==0.2
django-bcrypt==0.9.2
django-cors-headers==3.1.1
django-daterange-filter==1.3.0
django-dynamic-fixture==2.0.0
django-extensions==2.2.5
django-jsonfield==1.3.1
django-redis==4.10.0
django-redis-sessions==0.6.1
django-rest-auth==0.9.5
django-simple-history==2.7.3
django-suit==0.2.28
djangorestframework==3.10.3
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.3.0
django-mysql
mysqlclient==1.4.6
idna==2.8
isort==4.3.21
Pillow==6.2.1
py-bcrypt==0.4
python-dateutil==2.8.1
pytz==2019.3
redis==3.0.1
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3
uwsgi
newrelic
sentry-sdk==0.14.4
redis-py-cluster==2.0.0
